+++ changelog.md
@@ -1,9 +1,10 @@
 
 ### Version 0.1.0-SNAPSHOT
+ - [feature] github service / jcabi access setup
  - [feature] better test output for gradle
  - [feature] Added checkstyle
  - [feature] server port and log configuration
  - [feature] project setup
  - [initial] initial commit
 
--- Last change by: hschuetz <henning.schuetz@devbliss.com> Thu Feb 05 18:02:20 CET 2015
+-- Last change by: hschuetz <henning.schuetz@devbliss.com> Mon Feb 09 15:57:11 CET 2015